Change max_size and delay_seconds type to size_t Add MultithreadTest

+++ irohad/ordering/impl/ordering_service_impl.hpp
@@ -27,15 +27,15 @@ namespace iroha {
 
   class OrderingServiceImpl : public OrderingService{
    public:
-    explicit OrderingServiceImpl(uint16_t max_size, uint16_t delay_seconds);
+    explicit OrderingServiceImpl(size_t max_size, size_t delay_seconds);
     void propagate_transaction(const model::Transaction &transaction) override;
     rxcpp::observable<model::Proposal> on_proposal() override;
 
     void generateProposal();
    private:
     tbb::concurrent_queue<model::Transaction> queue_;
-    const uint16_t max_size_; // max number of txs in proposal
-    const uint16_t delay_seconds_; // wait for specified time if queue is empty
+    const size_t max_size_; // max number of txs in proposal
+    const size_t delay_seconds_; // wait for specified time if queue is empty
     rxcpp::subjects::subject<model::Proposal> proposals_;
   };
